+++ jest.setup.js
@@ -1,6 +1,9 @@
 import { configure, shallow, render, mount } from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
 
+// Make sure px object exists
+window.px = window.px || {};
+
 // React 16 Enzyme adapter
 configure({ adapter: new Adapter() });