+++ src/px-script/dialog/index.js
@@ -52,15 +52,17 @@ const dialog = (() => {
     const init = () => {
         const dialogEls = [...document.querySelectorAll(SELECTORS.DIALOG)];
 
-        if (dialogEls) {
-            const dialogs = dialogEls.map(dialog => new Dialog(dialog));
+        window.px._dialogs = window.px._dialogs || [];
+
+        if (dialogEls.length) {
+            dialogEls.forEach(dialog => window.px._dialogs.push(new Dialog(dialog)));
 
             // Init open buttons
             document.querySelectorAll(SELECTORS.OPEN).forEach(btn => {
                 const id = btn.dataset.dialogOpen;
                 let dialog;
 
-                dialogs.forEach(d => d.id === id ? dialog = d : null);
+                window.px._dialogs.forEach(d => d.id === id ? dialog = d : null);
 
                 if (dialog) {
                     btn.addEventListener("click", e => {
@@ -68,7 +70,7 @@ const dialog = (() => {
                         dialog.open();
                     });
                 } else {
-                    console.warn(`OPEN: No dialog with with id ${id} was found, make sure the attribute data-dialog-open contains the correct id.`);
+                    console.warn(`OPEN: No dialog with with id "${id}" was found, make sure the attribute data-dialog-open contains the correct id.`);
                 }
             });
 
@@ -77,7 +79,7 @@ const dialog = (() => {
                 const id = btn.dataset.dialogClose;
                 let dialog;
 
-                dialogs.forEach(d => d.id === id ? dialog = d : null);
+                window.px._dialogs.forEach(d => d.id === id ? dialog = d : null);
 
                 if (dialog) {
                     btn.addEventListener("click", e => {
@@ -85,13 +87,49 @@ const dialog = (() => {
                         dialog.close();
                     });
                 } else {
-                    console.warn(`CLOSE: No dialog with with id ${id} was found, make sure the attribute data-dialog-close contains the correct id.`);
+                    console.warn(`CLOSE: No dialog with with id "${id}" was found, make sure the attribute data-dialog-close contains the correct id.`);
                 }
             });
         }
     };
 
-    return { init };
+    const close = id => {
+        let dialog = null;
+
+        window.px._dialogs.forEach(d => d.id === id ? dialog = d : null);
+
+        try {
+            dialog.close();
+        } catch (e) {
+            console.error(`dialog.close: No dialog with id "${id}" found.`);
+
+            return false;
+        }
+
+        return dialog;
+    };
+
+    const open = id => {
+        let dialog = null;
+
+        window.px._dialogs.forEach(d => d.id === id ? dialog = d : null);
+
+        try {
+            dialog.open();
+        } catch (e) {
+            console.error(`dialog.open: No dialog with id "${id}" found.`);
+
+            return false;
+        }
+
+        return dialog;
+    };
+
+    return {
+        init,
+        close,
+        open
+    };
 })();
 
 export default dialog;